backend/requirements.txt
# Backend API requirements (lightweight)
Flask>=2.0.1
flask-cors>=4.0.0
python-dotenv==1.0.1
redis>=5.0.0
gunicorn>=21.2.0
